installed tensorboard for visuals and training monitoring
