+++ .gitignore
@@ -79,12 +79,6 @@ buck-out/
 .DS_Store
 
 .env
-AI/.ipynb_checkpoints/testingNeuralNets-checkpoint.ipynb
-AI/supervisedNeuralNet/.ipynb_checkpoints
-AI.supervisedNeuralNet/models/*
-AI/supervisedNeuralNet/supervisedNeuralNet/models/model/keras_metadata.pb
-AI/supervisedNeuralNet/supervisedNeuralNet/models/model/saved_model.pb
-AI/supervisedNeuralNet/supervisedNeuralNet/models/model/variables/variables.data-00000-of-00001
-AI/supervisedNeuralNet/supervisedNeuralNet/models/model/variables/variables.index
-AI/supervisedNeuralNet/model.h5
-AI/supervisedNeuralNet/testModel.h5
+
+AI/imageClassification/models/*/*.h5
+AI/imageClassification/models/*/logs/
